requirements.txt
fastapi==0.103.1
requests==2.26.0
PyYaml==6.0
uuid==1.30
uvicorn==0.22.0
websockets==10.3